dvc_checker/requirements.txt
pytest
python-consul
retrying
termcolor
# ipdb not working on gg tb
#ipdb
